ecr: Remove regional STS patch preferring updated AWS SDK.
The patch being removed changed the how the kubelet would locate and
acquire credentials for authenticating against ECR. It would replace the
built in, and default credential location behaviors with a custom path,
and an undocumented set of environment variables. With the patch to
upgrade the kubelet to version 1.34.24 of the AWS SDK, it looks like the
ability to make use of regional STS is possible via the
`sts_regional_endpoints` setting or the `AWS_STS_REGIONAL_ENDPOINTS`
environment variable.

The removal of this patch restores the expected and common credential
location behavior of the AWS SDK and hopefully remove the undocumented
environment variables from being needed.

Further context on this change can be found in
https://github.com/aws/eks-distro/issues/129